Modernized Android unit tests. Moved TestUtils into correct directory for its package. Removed redundant tests. Reformatted some code and renamed methods for clarity.
