+++ database/build.gradle
@@ -10,6 +10,7 @@ android {
         targetSdkVersion targetSdk
         versionCode 1
         versionName "1.0"
+        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
     }
 
     buildTypes {
